+++ app/[locale]/(authorization)/product/[id]/page.tsx
