+++ app/[locale]/(authorization)/singleBlog/[id]/page.tsx
