+++ app/[locale]/contants.ts
