+++ app/[locale]/interface.ts
