+++ app/[locale]/layout.tsx
@@ -1,0 +1,34 @@
+import Providers from "./providers";
+import "./globals.css";
+// import { RootLayoutProps } from "./interface";
+import React, { ReactElement } from "react";
+import { I18nProviderClient } from "../../locales/client";
+
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+interface RootLayoutProps {
+  params: {
+    locale: string;
+  };
+  children: ReactElement;
+}
+
+const RootLayout: React.FC<RootLayoutProps> = ({
+  params: { locale },
+  children,
+}: RootLayoutProps) => {
+  return (
+    <html lang="en" suppressHydrationWarning>
+      <body className="w-full h-full bg-[#ffffff] dark:bg-[#0C1B31] flex item-center justify-center transition-colors duration-700">
+        <I18nProviderClient locale={locale}>
+          <Providers>{children}</Providers>
+        </I18nProviderClient>
+      </body>
+    </html>
+  );
+};
+
+export default RootLayout;
