+++ components/mainContent/ProductItems.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import { ProductData } from "../../app/interface";
+import { ProductData } from "../../app/[locale]/interface";
 
 const ProductItems = ({
   imgUrl,
